Extract portfolio record evidence view model
